third_party/WebKit/LayoutTests/editing/deleting/delete-at-start-or-end.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div contenteditable>|</div>',
    'delete',
    '<div contenteditable>|</div>'),
    'Backward delete in an empty editable DIV');

test(() => assert_selection(
    '<div contenteditable>|</div>',
    'forwardDelete',
    '<div contenteditable>|</div>'),
    'Forward delete in an empty editable DIV');
</script>
